+++ cookbooks/devel/definitions/bootstrap.rb
@@ -3,7 +3,7 @@
 # Cookbook Name:: devel
 # Definitions:: bootstrap
 #
-# Copyright 2013, Carles Muiños
+# Copyright 2013,2014 Carles Muiños
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -23,77 +23,77 @@ define :bootstrap, :shell => "zsh", :folders => [], :files => [], :templates => 
   required = params[:requires] || "package[#{params[:name]}]"
 
   box = node[:box]
-  devel = box['devel']
+  devel = box[:devel]
 
-  devel['users'].each do |username|
-    usr = box['users'][username]
+  devel[:users].each do |username|
+    usr = box[:users][username]
 
     if params[:before]
       bash "bootstrap_#{params[:name]}_for_#{username}" do
         code params[:before]
         user username
-        group usr['group']
+        group usr[:group]
         action :nothing
         subscribes :run, resources(required), :immediately
       end
     end
 
     params[:folders].each do |folder|
-      dir = "#{usr['home']}/#{folder}"
+      dir = "#{usr[:home]}/#{folder}"
 
       directory_tree dir do
-        exclude usr['home']
+        exclude usr[:home]
         owner username
-        group usr['group']
+        group usr[:group]
         mode 00755
       end
     end
 
     params[:files].each do |f_|
-      f = "#{usr['home']}/#{f_}"
+      f = "#{usr[:home]}/#{f_}"
 
       devfile params[:name] do
         file f
         user username
-        group usr['group']
+        group usr[:group]
       end
     end
 
     params[:templates].each do |h|
-      file = "#{usr['home']}/#{h[:file]}"
-      vars = h[:vars].merge(:devel_dir => "#{usr['home']}/#{devel['folder']}")
+      file = "#{usr[:home]}/#{h[:file]}"
+      vars = h[:vars].merge(:devel_dir => "#{usr[:home]}/#{devel[:folder]}")
 
       devfile params[:name] do
         template file
         variables vars
         user username
-        group usr['group']
+        group usr[:group]
       end
     end
 
     params[:links].each do |h|
-      link "#{usr['home']}/#{h[:from]}" do
-        to "#{usr['home']}/#{h[:to]}"
+      link "#{usr[:home]}/#{h[:from]}" do
+        to "#{usr[:home]}/#{h[:to]}"
         owner username
-        group usr['group']
+        group usr[:group]
       end
     end
 
     if params[:aliases]
       aliases params[:name] do
-        dotfiles_dir "#{usr['home']}/#{box['dotfiles']['folder']}"
+        dotfiles_dir "#{usr[:home]}/#{box[:dotfiles][:folder]}"
         shell "zsh"
         user username
-        group usr['group']
+        group usr[:group]
       end
     end
 
     if params[:functions]
       functions params[:name] do
-        dotfiles_dir "#{usr['home']}/#{box['dotfiles']['folder']}"
+        dotfiles_dir "#{usr[:home]}/#{box[:dotfiles][:folder]}"
         shell "zsh"
         user username
-        group usr['group']
+        group usr[:group]
       end
     end
   end